Revert "Backend/feat/refactor member (#7)"
This reverts commit 69ab4d25c9708834a972fcfe2da9155b9a46542e.
