+++ backend/demo/src/main/java/com/_2cha/demo/member/controller/MemberController.java
@@ -12,8 +12,6 @@ import com._2cha.demo.member.dto.ToggleAchievementExposureResponse;
 import com._2cha.demo.member.service.MemberService;
 import jakarta.validation.Valid;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -48,10 +46,7 @@ public class MemberController {
   public List<ToggleAchievementExposureResponse> toggleAchvsExposure(@Authed Long memberId,
                                                                      @Valid @RequestBody List<ToggleAchievementExposureRequest> dto
                                                                     ) {
-
-    Map<Long, Boolean> achvExposureMap = dto.stream().collect(
-        Collectors.toMap(d -> d.getAchvId(), d -> d.isExposure()));
-    return memberService.toggleAchievementsExposure(memberId, achvExposureMap);
+    return memberService.toggleAchievementsExposure(memberId, dto);
   }
 
   @Auth
@@ -75,8 +70,8 @@ public class MemberController {
 
 
   @PostMapping
-  public MemberInfoResponse signUp(@Valid @RequestBody SignUpRequest dto) {
-    return memberService.signUp(dto.getEmail(), dto.getName(), dto.getPassword());
+  public MemberInfoResponse signUp(@Valid @RequestBody SignUpRequest dto) throws Exception {
+    return memberService.signUp(dto);
   }
 
   @Auth
